requirements.txt
# Core Framework
Django
djangorestframework
django-filter

# Database & ORM
psycopg2-binary
dj-database-url
supabase

# Authentication & Security
djangorestframework-simplejwt
cryptography
django-environ
PyJWT

# Ethereum Blockchain Integration
web3
eth-account
eth-utils
py-solc-x

# AI Integration - Google Gemini
google-generativeai
Pillow
# USSD Integration - Africa's Talking
africastalking

# Task Queue & Caching
celery
redis
django-redis
django-celery-beat

# Weather API
requests

# File Storage
boto3
django-storages

# Utilities
python-decouple
python-dotenv
phonenumbers
geopy
# Monitoring & Logging
sentry-sdk

# Testing
pytest
pytest-django
factory-boy
faker

# API Documentation
drf-spectacular

# Payment Integration (Mock for hackathon)
stripe

# Excel/CSV Export
openpyxl
pandas

# SMS & Email
twilio
sendgrid

# Rate Limiting
django-ratelimit

# Image Processing
opencv-python-headless

# Environment
gunicorn
whitenoise

stripe

deepgram-sdk